% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.clone.colours.in.order}
\alias{get.clone.colours.in.order}
\title{Generate a named vector of colors for every clone ID specified,
ordered by the clone IDs in `clone.order`.}
\usage{
get.clone.colours.in.order(clone.colours, clone.ids, clone.order = NULL)
}
\arguments{
\item{clone.colours}{A vector of colors assigned to clones. If `NULL`, colors will be generated automatically.}

\item{clone.ids}{A vector of clone identifiers.}

\item{clone.order}{An optional vector specifying the order of clones. If `NULL`, clone order is not gauranteed.}
}
\value{
A list containing:
\describe{
  \item{clone.colours}{A named vector of colors assigned to each clone.}
  \item{clone.order}{The ordered clones.}
}
}
\description{
Assigns colors to clones and ensures they follow a specified order.
Any colors specified in `clone.colours` are maintained in the order
specified and are used as the first colors for the `clone.ids`.
}
\details{
if `clone.colours` is NULL or an empty vector, _and_
`minimum.number.of.colours` is 0, `NULL` is returned
so BPG's default color selection can be used without needing to
check for `NULL`.
}
