% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.clone.colours}
\alias{get.clone.colours}
\title{Generate a named vector of colors for every clone ID specified.}
\usage{
get.clone.colours(clone.colours, clone.ids, minimum.number.of.colours = 0)
}
\arguments{
\item{clone.colours}{A vector of colors assigned to clones. If `NULL`, colors will be generated automatically.}

\item{clone.ids}{A vector of clone identifiers.}

\item{minimum.number.of.colours}{An integer specifying the minimum number of colors required.}
}
\value{
A named vector of colors assigned to each clone.
}
\description{
Assigns colors to clones, ensuring a minimum number of colors are used.
Any colors specified in `clone.colours` are maintained in the order
specified and are used as the first colors for the `clone.ids`.
}
\details{
if `clone.colours` is NULL or an empty vector, _and_
`minimum.number.of.colours` is 0, `NULL` is returned
so BPG's default color selection can be used without needing to
check for `NULL`.
}
