% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.colours.R
\name{get.colours.in.order}
\alias{get.colours.in.order}
\title{Generate a named vector of colors for every value specified,
ordered by the value in `value.order`.}
\usage{
get.colours.in.order(
  value.list,
  value.order = NULL,
  predetermined.colours = NULL
)
}
\arguments{
\item{value.list}{A vector of values.}

\item{value.order}{An optional vector specifying the order of values. If `NULL`, value order is not gauranteed.}

\item{predetermined.colours}{A vector of colors assigned to values. If `NULL`, colors will be generated automatically.}
}
\value{
A list containing:
\describe{
  \item{colours}{A named vector of colors assigned to each value.}
  \item{value.order}{The ordered values.}
}
}
\description{
Assigns colors to values and ensures they follow a specified order.
Any colors specified in `predetermined.colours` are maintained in the order
specified and are used as the first colors for the `value.list`.
}
