% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SRCGrob}
\alias{SRCGrob}
\title{Create a subclone plot}
\usage{
SRCGrob(
  ccf_df,
  tree_df,
  filename = "SRC_tree.pdf",
  scale1 = 0.05443424,
  scale2 = 0.5/362,
  wid = 1.2,
  line.lwd = 3,
  length_from_node_edge = TRUE,
  seg1.col = "black",
  seg2.col = "green",
  add_polygons = TRUE,
  extra_len = 10,
  sig_shape = 3,
  sig_curve = 3,
  spread = 1,
  fixed_angle = NULL,
  genes_df = NULL,
  genes = "default",
  gene_line_dist = 0.1,
  gene.cex = 0.85,
  add_genes = FALSE,
  genes_on_nodes = FALSE,
  yaxis_position = "left",
  yaxis1_label = "SNVs",
  yaxis2_label = NULL,
  axis_label_cex = list(x = 1.55, y = 1.55),
  axis_cex = list(x = 1.45, y = 1.45),
  yaxis1_interval = NA,
  yaxis2_interval = NA,
  ylimit = NULL,
  xaxis_label = NULL,
  xaxis_space_left = 0,
  xaxis_space_right = 0,
  min_width = NULL,
  nodes = "circle",
  rad = 0.1,
  label_nodes = TRUE,
  node_col = "grey29",
  label_cex = NA,
  cluster_list = NULL,
  add_normal = FALSE,
  normal_cex = 1,
  title = NULL,
  title.cex = 1.7,
  title.y = NULL,
  title.y.units = "npc",
  ...
)
}
\arguments{
\item{ccf_df}{}

\item{tree_df}{}

\item{filename}{}

\item{scale1}{}

\item{scale2}{}

\item{wid}{}

\item{line.lwd}{}

\item{length_from_node_edge}{}

\item{seg1.col}{}

\item{seg2.col}{}

\item{add_polygons}{}

\item{extra_len}{}

\item{sig_shape}{}

\item{sig_curve}{}

\item{spread}{}

\item{fixed_angle}{}

\item{genes_df}{}

\item{genes}{}

\item{gene_line_dist}{}

\item{gene.cex}{}

\item{add_genes}{}

\item{genes_on_nodes}{}

\item{yaxis_position}{}

\item{yaxis1_label}{}

\item{yaxis2_label}{}

\item{axis_label_cex}{}

\item{axis_cex}{}

\item{yaxis1_interval}{}

\item{yaxis2_interval}{}

\item{ylimit}{}

\item{xaxis_label}{}

\item{xaxis_space_left}{}

\item{xaxis_space_right}{}

\item{min_width}{}

\item{nodes}{}

\item{rad}{}

\item{label_nodes}{}

\item{node_col}{}

\item{label_cex}{}

\item{cluster_list}{}

\item{add_normal}{}

\item{normal_cex}{}

\item{title}{}

\item{title.cex}{}

\item{title.y}{}

\item{title.y.units}{}

\item{...}{}
}
\value{

}
\description{
Create a subclone plot
}
